Refs 22893: Refatored the test for the new feature
