+++ test/unittest/dds/xtypes/serializers/idl/DynTypeIDLTests.cpp
@@ -45,8 +45,7 @@ protected:
     {
         // Find TypeObjects for the type
         xtypes::TypeObjectPair type_objs;
-        ASSERT_EQ(DomainParticipantFactory::get_instance()->type_object_registry().get_type_objects(
-                    type_name, type_objs),
+        ASSERT_EQ(DomainParticipantFactory::get_instance()->type_object_registry().get_type_objects(type_name, type_objs),
                 fastdds::dds::RETCODE_OK);
 
         // Create DynamicType from TypeObject
@@ -100,52 +99,6 @@ TEST_P(DynTypeIDLTests, to_idl)
 
     // Read the IDL file as a string
     const auto file_name = std::string("types/") + type + "/" + type + ".idl";
-
-    std::ifstream file(file_name);
-    ASSERT_TRUE(file.is_open());
-
-    const std::string idl_file{std::istreambuf_iterator<char>(file), std::istreambuf_iterator<char>()};
-
-    // Get Dynamic type
-    DynamicType::_ref_type dyn_type;
-    get_dynamic_type(snake_to_camel(type), dyn_type);
-    ASSERT_NE(dyn_type, nullptr);
-
-    // Serialize DynamicType to IDL
-    std::stringstream idl_serialization;
-    ASSERT_EQ(idl_serialize(dyn_type, idl_serialization), RETCODE_OK);
-
-    // Compare IDLs
-    ASSERT_EQ(idl_file, idl_serialization.str());
-}
-
-/**
- * Verify that the IDL serialization of a DynamicType generated with Fast-DDS Gen matches its IDL file, with modules.
- *
- * CASES:
- *  - Verify that the IDL file was opened successfully.
- *  - Verify that the IDL serialization finished successfully.
- *  - Verify that the two IDLs match.
- */
-TEST_P(DynTypeIDLTests, to_idl_modules)
-{
-    const std::string type = GetParam();
-
-    if (type != "alias_struct" &&
-        type != "bitmask_struct" &&
-        type != "bitset_struct" &&
-        type != "enum_struct" &&
-        type != "struct_struct" &&
-        type != "union_struct")
-    {
-        GTEST_SKIP() << "Module structure only supported for alias, bitmask, bitset, enum, struct and union types.";
-        return;
-    }
-
-    test::register_type_object_representation(type);
-
-    // Read the IDL file as a string
-    const auto file_name = std::string("types/module_idls/") + type + "/" + type + ".idl";
 
     std::ifstream file(file_name);
     ASSERT_TRUE(file.is_open());
