+++ test/unittest/dds/xtypes/serializers/idl/types/all_types.hpp
@@ -47,6 +47,9 @@
 #include "map_struct/gen/map_struct.hpp"
 #include "map_struct/gen/map_structPubSubTypes.hpp"
 
+#include "module_struct/gen/module_struct.hpp"
+#include "module_struct/gen/module_structPubSubTypes.hpp"
+
 #include "primitives_struct/gen/primitives_struct.hpp"
 #include "primitives_struct/gen/primitives_structPubSubTypes.hpp"
 
@@ -74,6 +77,7 @@ const std::string ENUM_STRUCT{"enum_struct"};
 const std::string EXTENSIBILITY_STRUCT{"extensibility_struct"};
 const std::string KEY_STRUCT{"key_struct"};
 const std::string MAP_STRUCT{"map_struct"};
+const std::string MODULE_STRUCT{"module_struct"};
 const std::string PRIMITIVE_STRUCT{"primitives_struct"};
 const std::string SEQUENCE_STRUCT{"sequence_struct"};
 const std::string STRING_STRUCT{"string_struct"};
@@ -91,6 +95,7 @@ const std::vector<std::string> supported_types = {
     SupportedTypes::EXTENSIBILITY_STRUCT,
     SupportedTypes::KEY_STRUCT,
     SupportedTypes::MAP_STRUCT,
+    SupportedTypes::MODULE_STRUCT,
     SupportedTypes::PRIMITIVE_STRUCT,
     SupportedTypes::SEQUENCE_STRUCT,
     SupportedTypes::STRING_STRUCT,
@@ -136,6 +141,10 @@ void register_type_object_representation(
     else if (type_name == SupportedTypes::MAP_STRUCT)
     {
         type_support.reset(new MapStructPubSubType());
+    }
+    else if (type_name == SupportedTypes::MODULE_STRUCT)
+    {
+        type_support.reset(new ModuleStructPubSubType());
     }
     else if (type_name == SupportedTypes::PRIMITIVE_STRUCT)
     {
