--- test/unittest/dds/xtypes/serializers/idl/types/module_idls/alias_struct/alias_struct.idl
@@ -1,12 +1,0 @@
-module TrialModule
-{
-    module AliasModule
-    {
-        typedef long AliasType;
-    };
-};
-
-struct TrialStruct
-{
-    TrialModule::AliasModule::AliasType alias_member;
-};
