--- test/unittest/dds/xtypes/serializers/idl/types/module_idls/alias_struct/gen/alias_struct.hpp
@@ -1,188 +1,0 @@
-// Copyright 2016 Proyectos y Sistemas de Mantenimiento SL (eProsima).
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-/*!
- * @file alias_struct.hpp
- * This header file contains the declaration of the described types in the IDL file.
- *
- * This file was generated by the tool fastddsgen.
- */
-
-#ifndef FAST_DDS_GENERATED__ALIAS_STRUCT_HPP
-#define FAST_DDS_GENERATED__ALIAS_STRUCT_HPP
-
-#include <cstdint>
-#include <utility>
-
-#if defined(_WIN32)
-#if defined(EPROSIMA_USER_DLL_EXPORT)
-#define eProsima_user_DllExport __declspec( dllexport )
-#else
-#define eProsima_user_DllExport
-#endif  // EPROSIMA_USER_DLL_EXPORT
-#else
-#define eProsima_user_DllExport
-#endif  // _WIN32
-
-#if defined(_WIN32)
-#if defined(EPROSIMA_USER_DLL_EXPORT)
-#if defined(ALIAS_STRUCT_SOURCE)
-#define ALIAS_STRUCT_DllAPI __declspec( dllexport )
-#else
-#define ALIAS_STRUCT_DllAPI __declspec( dllimport )
-#endif // ALIAS_STRUCT_SOURCE
-#else
-#define ALIAS_STRUCT_DllAPI
-#endif  // EPROSIMA_USER_DLL_EXPORT
-#else
-#define ALIAS_STRUCT_DllAPI
-#endif // _WIN32
-
-namespace TrialModule {
-
-namespace AliasModule {
-
-typedef int32_t AliasType;
-
-
-} // namespace AliasModule
-
-} // namespace TrialModule
-/*!
- * @brief This class represents the structure TrialStruct defined by the user in the IDL file.
- * @ingroup alias_struct
- */
-class TrialStruct
-{
-public:
-
-    /*!
-     * @brief Default constructor.
-     */
-    eProsima_user_DllExport TrialStruct()
-    {
-    }
-
-    /*!
-     * @brief Default destructor.
-     */
-    eProsima_user_DllExport ~TrialStruct()
-    {
-    }
-
-    /*!
-     * @brief Copy constructor.
-     * @param x Reference to the object TrialStruct that will be copied.
-     */
-    eProsima_user_DllExport TrialStruct(
-            const TrialStruct& x)
-    {
-                    m_alias_member = x.m_alias_member;
-
-    }
-
-    /*!
-     * @brief Move constructor.
-     * @param x Reference to the object TrialStruct that will be copied.
-     */
-    eProsima_user_DllExport TrialStruct(
-            TrialStruct&& x) noexcept
-    {
-        m_alias_member = x.m_alias_member;
-    }
-
-    /*!
-     * @brief Copy assignment.
-     * @param x Reference to the object TrialStruct that will be copied.
-     */
-    eProsima_user_DllExport TrialStruct& operator =(
-            const TrialStruct& x)
-    {
-
-                    m_alias_member = x.m_alias_member;
-
-        return *this;
-    }
-
-    /*!
-     * @brief Move assignment.
-     * @param x Reference to the object TrialStruct that will be copied.
-     */
-    eProsima_user_DllExport TrialStruct& operator =(
-            TrialStruct&& x) noexcept
-    {
-
-        m_alias_member = x.m_alias_member;
-        return *this;
-    }
-
-    /*!
-     * @brief Comparison operator.
-     * @param x TrialStruct object to compare.
-     */
-    eProsima_user_DllExport bool operator ==(
-            const TrialStruct& x) const
-    {
-        return (m_alias_member == x.m_alias_member);
-    }
-
-    /*!
-     * @brief Comparison operator.
-     * @param x TrialStruct object to compare.
-     */
-    eProsima_user_DllExport bool operator !=(
-            const TrialStruct& x) const
-    {
-        return !(*this == x);
-    }
-
-    /*!
-     * @brief This function sets a value in member alias_member
-     * @param _alias_member New value for member alias_member
-     */
-    eProsima_user_DllExport void alias_member(
-            TrialModule::AliasModule::AliasType _alias_member)
-    {
-        m_alias_member = _alias_member;
-    }
-
-    /*!
-     * @brief This function returns the value of member alias_member
-     * @return Value of member alias_member
-     */
-    eProsima_user_DllExport TrialModule::AliasModule::AliasType alias_member() const
-    {
-        return m_alias_member;
-    }
-
-    /*!
-     * @brief This function returns a reference to member alias_member
-     * @return Reference to member alias_member
-     */
-    eProsima_user_DllExport TrialModule::AliasModule::AliasType& alias_member()
-    {
-        return m_alias_member;
-    }
-
-
-
-private:
-
-    TrialModule::AliasModule::AliasType m_alias_member{0};
-
-};
-
-#endif // _FAST_DDS_GENERATED_ALIAS_STRUCT_HPP_
-
-
