--- test/unittest/dds/xtypes/serializers/idl/types/module_idls/bitmask_struct/bitmask_struct.idl
@@ -1,18 +1,0 @@
-module TrialModule
-{
-    module BitmaskModule
-    {
-        @bit_bound(8)
-        bitmask BitmaskType
-        {
-            FLAG_ONE,
-            FLAG_TWO,
-            FLAG_THREE
-        };
-    };
-};
-
-struct TrialStruct
-{
-    TrialModule::BitmaskModule::BitmaskType bitmask_member;
-};
