--- test/unittest/dds/xtypes/serializers/idl/types/module_idls/bitmask_struct/gen/bitmask_struct.hpp
@@ -1,207 +1,0 @@
-// Copyright 2016 Proyectos y Sistemas de Mantenimiento SL (eProsima).
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-/*!
- * @file bitmask_struct.hpp
- * This header file contains the declaration of the described types in the IDL file.
- *
- * This file was generated by the tool fastddsgen.
- */
-
-#ifndef FAST_DDS_GENERATED__BITMASK_STRUCT_HPP
-#define FAST_DDS_GENERATED__BITMASK_STRUCT_HPP
-
-#include <cstdint>
-#include <utility>
-
-#if defined(_WIN32)
-#if defined(EPROSIMA_USER_DLL_EXPORT)
-#define eProsima_user_DllExport __declspec( dllexport )
-#else
-#define eProsima_user_DllExport
-#endif  // EPROSIMA_USER_DLL_EXPORT
-#else
-#define eProsima_user_DllExport
-#endif  // _WIN32
-
-#if defined(_WIN32)
-#if defined(EPROSIMA_USER_DLL_EXPORT)
-#if defined(BITMASK_STRUCT_SOURCE)
-#define BITMASK_STRUCT_DllAPI __declspec( dllexport )
-#else
-#define BITMASK_STRUCT_DllAPI __declspec( dllimport )
-#endif // BITMASK_STRUCT_SOURCE
-#else
-#define BITMASK_STRUCT_DllAPI
-#endif  // EPROSIMA_USER_DLL_EXPORT
-#else
-#define BITMASK_STRUCT_DllAPI
-#endif // _WIN32
-
-namespace TrialModule {
-
-namespace BitmaskModule {
-
-/*!
- * @brief This enumeration represents the BitmaskType bitflags defined by the user in the IDL file.
- * @ingroup bitmask_struct
- */
-enum BitmaskTypeBits : uint8_t
-{
-    FLAG_ONE = 0x01ull << 0,
-    FLAG_TWO = 0x01ull << 1,
-    FLAG_THREE = 0x01ull << 2
-};
-typedef uint8_t BitmaskType;
-
-} // namespace BitmaskModule
-
-} // namespace TrialModule
-/*!
- * @brief This class represents the structure TrialStruct defined by the user in the IDL file.
- * @ingroup bitmask_struct
- */
-class TrialStruct
-{
-public:
-
-    /*!
-     * @brief Default constructor.
-     */
-    eProsima_user_DllExport TrialStruct()
-    {
-    }
-
-    /*!
-     * @brief Default destructor.
-     */
-    eProsima_user_DllExport ~TrialStruct()
-    {
-    }
-
-    /*!
-     * @brief Copy constructor.
-     * @param x Reference to the object TrialStruct that will be copied.
-     */
-    eProsima_user_DllExport TrialStruct(
-            const TrialStruct& x)
-    {
-                    m_bitmask_member = x.m_bitmask_member;
-
-    }
-
-    /*!
-     * @brief Move constructor.
-     * @param x Reference to the object TrialStruct that will be copied.
-     */
-    eProsima_user_DllExport TrialStruct(
-            TrialStruct&& x) noexcept
-    {
-        m_bitmask_member = std::move(x.m_bitmask_member);
-    }
-
-    /*!
-     * @brief Copy assignment.
-     * @param x Reference to the object TrialStruct that will be copied.
-     */
-    eProsima_user_DllExport TrialStruct& operator =(
-            const TrialStruct& x)
-    {
-
-                    m_bitmask_member = x.m_bitmask_member;
-
-        return *this;
-    }
-
-    /*!
-     * @brief Move assignment.
-     * @param x Reference to the object TrialStruct that will be copied.
-     */
-    eProsima_user_DllExport TrialStruct& operator =(
-            TrialStruct&& x) noexcept
-    {
-
-        m_bitmask_member = std::move(x.m_bitmask_member);
-        return *this;
-    }
-
-    /*!
-     * @brief Comparison operator.
-     * @param x TrialStruct object to compare.
-     */
-    eProsima_user_DllExport bool operator ==(
-            const TrialStruct& x) const
-    {
-        return (m_bitmask_member == x.m_bitmask_member);
-    }
-
-    /*!
-     * @brief Comparison operator.
-     * @param x TrialStruct object to compare.
-     */
-    eProsima_user_DllExport bool operator !=(
-            const TrialStruct& x) const
-    {
-        return !(*this == x);
-    }
-
-    /*!
-     * @brief This function copies the value in member bitmask_member
-     * @param _bitmask_member New value to be copied in member bitmask_member
-     */
-    eProsima_user_DllExport void bitmask_member(
-            const TrialModule::BitmaskModule::BitmaskType& _bitmask_member)
-    {
-        m_bitmask_member = _bitmask_member;
-    }
-
-    /*!
-     * @brief This function moves the value in member bitmask_member
-     * @param _bitmask_member New value to be moved in member bitmask_member
-     */
-    eProsima_user_DllExport void bitmask_member(
-            TrialModule::BitmaskModule::BitmaskType&& _bitmask_member)
-    {
-        m_bitmask_member = std::move(_bitmask_member);
-    }
-
-    /*!
-     * @brief This function returns a constant reference to member bitmask_member
-     * @return Constant reference to member bitmask_member
-     */
-    eProsima_user_DllExport const TrialModule::BitmaskModule::BitmaskType& bitmask_member() const
-    {
-        return m_bitmask_member;
-    }
-
-    /*!
-     * @brief This function returns a reference to member bitmask_member
-     * @return Reference to member bitmask_member
-     */
-    eProsima_user_DllExport TrialModule::BitmaskModule::BitmaskType& bitmask_member()
-    {
-        return m_bitmask_member;
-    }
-
-
-
-private:
-
-    TrialModule::BitmaskModule::BitmaskType m_bitmask_member{0};
-
-};
-
-#endif // _FAST_DDS_GENERATED_BITMASK_STRUCT_HPP_
-
-
