--- test/unittest/dds/xtypes/serializers/idl/types/module_idls/bitmask_struct/gen/bitmask_structTypeObjectSupport.cxx
@@ -1,171 +1,0 @@
-// Copyright 2016 Proyectos y Sistemas de Mantenimiento SL (eProsima).
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-/*!
- * @file bitmask_structTypeObjectSupport.cxx
- * Source file containing the implementation to register the TypeObject representation of the described types in the IDL file
- *
- * This file was generated by the tool fastddsgen.
- */
-
-#include "bitmask_structTypeObjectSupport.hpp"
-
-#include <mutex>
-#include <string>
-
-#include <fastcdr/xcdr/external.hpp>
-#include <fastcdr/xcdr/optional.hpp>
-#include <fastdds/dds/domain/DomainParticipantFactory.hpp>
-#include <fastdds/dds/log/Log.hpp>
-#include <fastdds/dds/xtypes/common.hpp>
-#include <fastdds/dds/xtypes/type_representation/ITypeObjectRegistry.hpp>
-#include <fastdds/dds/xtypes/type_representation/TypeObject.hpp>
-#include <fastdds/dds/xtypes/type_representation/TypeObjectUtils.hpp>
-
-#include "bitmask_struct.hpp"
-
-
-using namespace eprosima::fastdds::dds::xtypes;
-
-namespace TrialModule {
-namespace BitmaskModule {
-void register_BitmaskType_type_identifier(
-        TypeIdentifierPair& type_ids_BitmaskType)
-{
-    ReturnCode_t return_code_BitmaskType {eprosima::fastdds::dds::RETCODE_OK};
-    return_code_BitmaskType =
-        eprosima::fastdds::dds::DomainParticipantFactory::get_instance()->type_object_registry().get_type_identifiers(
-        "TrialModule::BitmaskModule::BitmaskType", type_ids_BitmaskType);
-    if (eprosima::fastdds::dds::RETCODE_OK != return_code_BitmaskType)
-    {
-        BitmaskTypeFlag bitmask_flags_BitmaskType = 0;
-        BitBound bit_bound_BitmaskType = 8;
-        CommonEnumeratedHeader common_BitmaskType = TypeObjectUtils::build_common_enumerated_header(bit_bound_BitmaskType, true);
-        QualifiedTypeName type_name_BitmaskType = "TrialModule::BitmaskModule::BitmaskType";
-        eprosima::fastcdr::optional<AppliedBuiltinTypeAnnotations> type_ann_builtin_BitmaskType;
-        eprosima::fastcdr::optional<AppliedAnnotationSeq> ann_custom_BitmaskType;
-        AppliedAnnotationSeq tmp_ann_custom_BitmaskType;
-        eprosima::fastcdr::optional<AppliedVerbatimAnnotation> verbatim_BitmaskType;
-        if (!tmp_ann_custom_BitmaskType.empty())
-        {
-            ann_custom_BitmaskType = tmp_ann_custom_BitmaskType;
-        }
-
-        CompleteTypeDetail detail_BitmaskType = TypeObjectUtils::build_complete_type_detail(type_ann_builtin_BitmaskType, ann_custom_BitmaskType, type_name_BitmaskType.to_string());
-        CompleteEnumeratedHeader header_BitmaskType = TypeObjectUtils::build_complete_enumerated_header(common_BitmaskType, detail_BitmaskType, true);
-        CompleteBitflagSeq flag_seq_BitmaskType;
-        {
-            uint16_t position_FLAG_ONE = 0;
-            BitflagFlag flags_FLAG_ONE = 0;
-            CommonBitflag common_FLAG_ONE = TypeObjectUtils::build_common_bitflag(position_FLAG_ONE, flags_FLAG_ONE);
-            eprosima::fastcdr::optional<AppliedBuiltinMemberAnnotations> member_ann_builtin_FLAG_ONE;
-            ann_custom_BitmaskType.reset();
-            MemberName name_FLAG_ONE = "FLAG_ONE";
-            CompleteMemberDetail detail_FLAG_ONE = TypeObjectUtils::build_complete_member_detail(name_FLAG_ONE, member_ann_builtin_FLAG_ONE, ann_custom_BitmaskType);
-            CompleteBitflag bitflag_FLAG_ONE = TypeObjectUtils::build_complete_bitflag(common_FLAG_ONE, detail_FLAG_ONE);
-            TypeObjectUtils::add_complete_bitflag(flag_seq_BitmaskType, bitflag_FLAG_ONE);
-        }
-        {
-            uint16_t position_FLAG_TWO = 1;
-            BitflagFlag flags_FLAG_TWO = 0;
-            CommonBitflag common_FLAG_TWO = TypeObjectUtils::build_common_bitflag(position_FLAG_TWO, flags_FLAG_TWO);
-            eprosima::fastcdr::optional<AppliedBuiltinMemberAnnotations> member_ann_builtin_FLAG_TWO;
-            ann_custom_BitmaskType.reset();
-            MemberName name_FLAG_TWO = "FLAG_TWO";
-            CompleteMemberDetail detail_FLAG_TWO = TypeObjectUtils::build_complete_member_detail(name_FLAG_TWO, member_ann_builtin_FLAG_TWO, ann_custom_BitmaskType);
-            CompleteBitflag bitflag_FLAG_TWO = TypeObjectUtils::build_complete_bitflag(common_FLAG_TWO, detail_FLAG_TWO);
-            TypeObjectUtils::add_complete_bitflag(flag_seq_BitmaskType, bitflag_FLAG_TWO);
-        }
-        {
-            uint16_t position_FLAG_THREE = 2;
-            BitflagFlag flags_FLAG_THREE = 0;
-            CommonBitflag common_FLAG_THREE = TypeObjectUtils::build_common_bitflag(position_FLAG_THREE, flags_FLAG_THREE);
-            eprosima::fastcdr::optional<AppliedBuiltinMemberAnnotations> member_ann_builtin_FLAG_THREE;
-            ann_custom_BitmaskType.reset();
-            MemberName name_FLAG_THREE = "FLAG_THREE";
-            CompleteMemberDetail detail_FLAG_THREE = TypeObjectUtils::build_complete_member_detail(name_FLAG_THREE, member_ann_builtin_FLAG_THREE, ann_custom_BitmaskType);
-            CompleteBitflag bitflag_FLAG_THREE = TypeObjectUtils::build_complete_bitflag(common_FLAG_THREE, detail_FLAG_THREE);
-            TypeObjectUtils::add_complete_bitflag(flag_seq_BitmaskType, bitflag_FLAG_THREE);
-        }
-        CompleteBitmaskType bitmask_type_BitmaskType = TypeObjectUtils::build_complete_bitmask_type(bitmask_flags_BitmaskType, header_BitmaskType, flag_seq_BitmaskType);
-        if (eprosima::fastdds::dds::RETCODE_BAD_PARAMETER ==
-                TypeObjectUtils::build_and_register_bitmask_type_object(bitmask_type_BitmaskType,
-                    type_name_BitmaskType.to_string(), type_ids_BitmaskType))
-        {
-            EPROSIMA_LOG_ERROR(XTYPES_TYPE_REPRESENTATION,
-                "TrialModule::BitmaskModule::BitmaskType already registered in TypeObjectRegistry for a different type.");
-        }
-    }
-}
-} // namespace BitmaskModule
-
-} // namespace TrialModule
-// TypeIdentifier is returned by reference: dependent structures/unions are registered in this same method
-void register_TrialStruct_type_identifier(
-        TypeIdentifierPair& type_ids_TrialStruct)
-{
-
-    ReturnCode_t return_code_TrialStruct {eprosima::fastdds::dds::RETCODE_OK};
-    return_code_TrialStruct =
-        eprosima::fastdds::dds::DomainParticipantFactory::get_instance()->type_object_registry().get_type_identifiers(
-        "TrialStruct", type_ids_TrialStruct);
-    if (eprosima::fastdds::dds::RETCODE_OK != return_code_TrialStruct)
-    {
-        StructTypeFlag struct_flags_TrialStruct = TypeObjectUtils::build_struct_type_flag(eprosima::fastdds::dds::xtypes::ExtensibilityKind::APPENDABLE,
-                false, false);
-        QualifiedTypeName type_name_TrialStruct = "TrialStruct";
-        eprosima::fastcdr::optional<AppliedBuiltinTypeAnnotations> type_ann_builtin_TrialStruct;
-        eprosima::fastcdr::optional<AppliedAnnotationSeq> ann_custom_TrialStruct;
-        CompleteTypeDetail detail_TrialStruct = TypeObjectUtils::build_complete_type_detail(type_ann_builtin_TrialStruct, ann_custom_TrialStruct, type_name_TrialStruct.to_string());
-        CompleteStructHeader header_TrialStruct;
-        header_TrialStruct = TypeObjectUtils::build_complete_struct_header(TypeIdentifier(), detail_TrialStruct);
-        CompleteStructMemberSeq member_seq_TrialStruct;
-        {
-            TypeIdentifierPair type_ids_bitmask_member;
-            ReturnCode_t return_code_bitmask_member {eprosima::fastdds::dds::RETCODE_OK};
-            return_code_bitmask_member =
-                eprosima::fastdds::dds::DomainParticipantFactory::get_instance()->type_object_registry().get_type_identifiers(
-                "TrialModule::BitmaskModule::BitmaskType", type_ids_bitmask_member);
-
-            if (eprosima::fastdds::dds::RETCODE_OK != return_code_bitmask_member)
-            {
-                TrialModule::BitmaskModule::register_BitmaskType_type_identifier(type_ids_bitmask_member);
-            }
-            StructMemberFlag member_flags_bitmask_member = TypeObjectUtils::build_struct_member_flag(eprosima::fastdds::dds::xtypes::TryConstructFailAction::DISCARD,
-                    false, false, false, false);
-            MemberId member_id_bitmask_member = 0x00000000;
-            bool common_bitmask_member_ec {false};
-            CommonStructMember common_bitmask_member {TypeObjectUtils::build_common_struct_member(member_id_bitmask_member, member_flags_bitmask_member, TypeObjectUtils::retrieve_complete_type_identifier(type_ids_bitmask_member, common_bitmask_member_ec))};
-            if (!common_bitmask_member_ec)
-            {
-                EPROSIMA_LOG_ERROR(XTYPES_TYPE_REPRESENTATION, "Structure bitmask_member member TypeIdentifier inconsistent.");
-                return;
-            }
-            MemberName name_bitmask_member = "bitmask_member";
-            eprosima::fastcdr::optional<AppliedBuiltinMemberAnnotations> member_ann_builtin_bitmask_member;
-            ann_custom_TrialStruct.reset();
-            CompleteMemberDetail detail_bitmask_member = TypeObjectUtils::build_complete_member_detail(name_bitmask_member, member_ann_builtin_bitmask_member, ann_custom_TrialStruct);
-            CompleteStructMember member_bitmask_member = TypeObjectUtils::build_complete_struct_member(common_bitmask_member, detail_bitmask_member);
-            TypeObjectUtils::add_complete_struct_member(member_seq_TrialStruct, member_bitmask_member);
-        }
-        CompleteStructType struct_type_TrialStruct = TypeObjectUtils::build_complete_struct_type(struct_flags_TrialStruct, header_TrialStruct, member_seq_TrialStruct);
-        if (eprosima::fastdds::dds::RETCODE_BAD_PARAMETER ==
-                TypeObjectUtils::build_and_register_struct_type_object(struct_type_TrialStruct, type_name_TrialStruct.to_string(), type_ids_TrialStruct))
-        {
-            EPROSIMA_LOG_ERROR(XTYPES_TYPE_REPRESENTATION,
-                    "TrialStruct already registered in TypeObjectRegistry for a different type.");
-        }
-    }
-}
-
