--- test/unittest/dds/xtypes/serializers/idl/types/module_idls/bitset_struct/bitset_struct.idl
@@ -1,17 +1,0 @@
-module TrialModule
-{
-    module BitsetModule
-    {
-        bitset BitsetType
-        {
-            bitfield<3> flag1;
-            bitfield<10> flag2;
-            bitfield<12, long> flag3;
-        };
-    };
-};
-
-struct TrialStruct
-{
-    TrialModule::BitsetModule::BitsetType bitset_member;
-};
