--- test/unittest/dds/xtypes/serializers/idl/types/module_idls/bitset_struct/gen/bitset_struct.hpp
@@ -1,231 +1,0 @@
-// Copyright 2016 Proyectos y Sistemas de Mantenimiento SL (eProsima).
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-/*!
- * @file bitset_struct.hpp
- * This header file contains the declaration of the described types in the IDL file.
- *
- * This file was generated by the tool fastddsgen.
- */
-
-#ifndef FAST_DDS_GENERATED__BITSET_STRUCT_HPP
-#define FAST_DDS_GENERATED__BITSET_STRUCT_HPP
-
-#include <cstdint>
-#include <utility>
-
-#if defined(_WIN32)
-#if defined(EPROSIMA_USER_DLL_EXPORT)
-#define eProsima_user_DllExport __declspec( dllexport )
-#else
-#define eProsima_user_DllExport
-#endif  // EPROSIMA_USER_DLL_EXPORT
-#else
-#define eProsima_user_DllExport
-#endif  // _WIN32
-
-#if defined(_WIN32)
-#if defined(EPROSIMA_USER_DLL_EXPORT)
-#if defined(BITSET_STRUCT_SOURCE)
-#define BITSET_STRUCT_DllAPI __declspec( dllexport )
-#else
-#define BITSET_STRUCT_DllAPI __declspec( dllimport )
-#endif // BITSET_STRUCT_SOURCE
-#else
-#define BITSET_STRUCT_DllAPI
-#endif  // EPROSIMA_USER_DLL_EXPORT
-#else
-#define BITSET_STRUCT_DllAPI
-#endif // _WIN32
-
-namespace TrialModule {
-
-namespace BitsetModule {
-
-/*!
- * @brief This structure represents the bitset BitsetType defined by the user in the IDL file.
- * @ingroup bitset_struct
- */
-struct BitsetType
-{
-        uint8_t flag1 : 3;
-
-        uint16_t flag2 : 10;
-
-        int32_t flag3 : 12;
-
-
-    /*!
-     * @brief Comparison operator.
-     * @param x BitsetType object to compare.
-     */
-    eProsima_user_DllExport bool operator ==(
-            const BitsetType& x) const
-    {
-        return (flag1 == x.flag1 &&
-           flag2 == x.flag2 &&
-           flag3 == x.flag3);
-    }
-
-    /*!
-     * @brief Comparison operator.
-     * @param x BitsetType object to compare.
-     */
-    eProsima_user_DllExport bool operator !=(
-            const BitsetType& x) const
-    {
-        return !(*this == x);
-    }
-};
-
-} // namespace BitsetModule
-
-} // namespace TrialModule
-/*!
- * @brief This class represents the structure TrialStruct defined by the user in the IDL file.
- * @ingroup bitset_struct
- */
-class TrialStruct
-{
-public:
-
-    /*!
-     * @brief Default constructor.
-     */
-    eProsima_user_DllExport TrialStruct()
-    {
-    }
-
-    /*!
-     * @brief Default destructor.
-     */
-    eProsima_user_DllExport ~TrialStruct()
-    {
-    }
-
-    /*!
-     * @brief Copy constructor.
-     * @param x Reference to the object TrialStruct that will be copied.
-     */
-    eProsima_user_DllExport TrialStruct(
-            const TrialStruct& x)
-    {
-                    m_bitset_member = x.m_bitset_member;
-
-    }
-
-    /*!
-     * @brief Move constructor.
-     * @param x Reference to the object TrialStruct that will be copied.
-     */
-    eProsima_user_DllExport TrialStruct(
-            TrialStruct&& x) noexcept
-    {
-        m_bitset_member = std::move(x.m_bitset_member);
-    }
-
-    /*!
-     * @brief Copy assignment.
-     * @param x Reference to the object TrialStruct that will be copied.
-     */
-    eProsima_user_DllExport TrialStruct& operator =(
-            const TrialStruct& x)
-    {
-
-                    m_bitset_member = x.m_bitset_member;
-
-        return *this;
-    }
-
-    /*!
-     * @brief Move assignment.
-     * @param x Reference to the object TrialStruct that will be copied.
-     */
-    eProsima_user_DllExport TrialStruct& operator =(
-            TrialStruct&& x) noexcept
-    {
-
-        m_bitset_member = std::move(x.m_bitset_member);
-        return *this;
-    }
-
-    /*!
-     * @brief Comparison operator.
-     * @param x TrialStruct object to compare.
-     */
-    eProsima_user_DllExport bool operator ==(
-            const TrialStruct& x) const
-    {
-        return (m_bitset_member == x.m_bitset_member);
-    }
-
-    /*!
-     * @brief Comparison operator.
-     * @param x TrialStruct object to compare.
-     */
-    eProsima_user_DllExport bool operator !=(
-            const TrialStruct& x) const
-    {
-        return !(*this == x);
-    }
-
-    /*!
-     * @brief This function copies the value in member bitset_member
-     * @param _bitset_member New value to be copied in member bitset_member
-     */
-    eProsima_user_DllExport void bitset_member(
-            const TrialModule::BitsetModule::BitsetType& _bitset_member)
-    {
-        m_bitset_member = _bitset_member;
-    }
-
-    /*!
-     * @brief This function moves the value in member bitset_member
-     * @param _bitset_member New value to be moved in member bitset_member
-     */
-    eProsima_user_DllExport void bitset_member(
-            TrialModule::BitsetModule::BitsetType&& _bitset_member)
-    {
-        m_bitset_member = std::move(_bitset_member);
-    }
-
-    /*!
-     * @brief This function returns a constant reference to member bitset_member
-     * @return Constant reference to member bitset_member
-     */
-    eProsima_user_DllExport const TrialModule::BitsetModule::BitsetType& bitset_member() const
-    {
-        return m_bitset_member;
-    }
-
-    /*!
-     * @brief This function returns a reference to member bitset_member
-     * @return Reference to member bitset_member
-     */
-    eProsima_user_DllExport TrialModule::BitsetModule::BitsetType& bitset_member()
-    {
-        return m_bitset_member;
-    }
-
-
-
-private:
-
-    TrialModule::BitsetModule::BitsetType m_bitset_member{};
-
-};
-
-#endif // _FAST_DDS_GENERATED_BITSET_STRUCT_HPP_
-
-
