--- test/unittest/dds/xtypes/serializers/idl/types/module_idls/enum_struct/enum_struct.idl
@@ -1,17 +1,0 @@
-module TrialModule
-{
-    module EnumModule
-    {
-        enum EnumType
-        {
-            VALUE_ONE,
-            VALUE_TWO,
-            VALUE_THREE
-        };
-    };
-};
-
-struct TrialStruct
-{
-    TrialModule::EnumModule::EnumType enum_member;
-};
