--- test/unittest/dds/xtypes/serializers/idl/types/module_idls/struct_struct/gen/struct_structTypeObjectSupport.cxx
@@ -1,171 +1,0 @@
-// Copyright 2016 Proyectos y Sistemas de Mantenimiento SL (eProsima).
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-/*!
- * @file struct_structTypeObjectSupport.cxx
- * Source file containing the implementation to register the TypeObject representation of the described types in the IDL file
- *
- * This file was generated by the tool fastddsgen.
- */
-
-#include "struct_structTypeObjectSupport.hpp"
-
-#include <mutex>
-#include <string>
-
-#include <fastcdr/xcdr/external.hpp>
-#include <fastcdr/xcdr/optional.hpp>
-#include <fastdds/dds/domain/DomainParticipantFactory.hpp>
-#include <fastdds/dds/log/Log.hpp>
-#include <fastdds/dds/xtypes/common.hpp>
-#include <fastdds/dds/xtypes/type_representation/ITypeObjectRegistry.hpp>
-#include <fastdds/dds/xtypes/type_representation/TypeObject.hpp>
-#include <fastdds/dds/xtypes/type_representation/TypeObjectUtils.hpp>
-
-#include "struct_struct.hpp"
-
-
-using namespace eprosima::fastdds::dds::xtypes;
-
-namespace TrialModule {
-namespace StructModule {
-// TypeIdentifier is returned by reference: dependent structures/unions are registered in this same method
-void register_StructType_type_identifier(
-        TypeIdentifierPair& type_ids_StructType)
-{
-
-    ReturnCode_t return_code_StructType {eprosima::fastdds::dds::RETCODE_OK};
-    return_code_StructType =
-        eprosima::fastdds::dds::DomainParticipantFactory::get_instance()->type_object_registry().get_type_identifiers(
-        "TrialModule::StructModule::StructType", type_ids_StructType);
-    if (eprosima::fastdds::dds::RETCODE_OK != return_code_StructType)
-    {
-        StructTypeFlag struct_flags_StructType = TypeObjectUtils::build_struct_type_flag(eprosima::fastdds::dds::xtypes::ExtensibilityKind::APPENDABLE,
-                false, false);
-        QualifiedTypeName type_name_StructType = "TrialModule::StructModule::StructType";
-        eprosima::fastcdr::optional<AppliedBuiltinTypeAnnotations> type_ann_builtin_StructType;
-        eprosima::fastcdr::optional<AppliedAnnotationSeq> ann_custom_StructType;
-        CompleteTypeDetail detail_StructType = TypeObjectUtils::build_complete_type_detail(type_ann_builtin_StructType, ann_custom_StructType, type_name_StructType.to_string());
-        CompleteStructHeader header_StructType;
-        header_StructType = TypeObjectUtils::build_complete_struct_header(TypeIdentifier(), detail_StructType);
-        CompleteStructMemberSeq member_seq_StructType;
-        {
-            TypeIdentifierPair type_ids_struct_data;
-            ReturnCode_t return_code_struct_data {eprosima::fastdds::dds::RETCODE_OK};
-            return_code_struct_data =
-                eprosima::fastdds::dds::DomainParticipantFactory::get_instance()->type_object_registry().get_type_identifiers(
-                "anonymous_string_unbounded", type_ids_struct_data);
-
-            if (eprosima::fastdds::dds::RETCODE_OK != return_code_struct_data)
-            {
-                {
-                    SBound bound = 0;
-                    StringSTypeDefn string_sdefn = TypeObjectUtils::build_string_s_type_defn(bound);
-                    if (eprosima::fastdds::dds::RETCODE_BAD_PARAMETER ==
-                            TypeObjectUtils::build_and_register_s_string_type_identifier(string_sdefn,
-                            "anonymous_string_unbounded", type_ids_struct_data))
-                    {
-                        EPROSIMA_LOG_ERROR(XTYPES_TYPE_REPRESENTATION,
-                            "anonymous_string_unbounded already registered in TypeObjectRegistry for a different type.");
-                    }
-                }
-            }
-            StructMemberFlag member_flags_struct_data = TypeObjectUtils::build_struct_member_flag(eprosima::fastdds::dds::xtypes::TryConstructFailAction::DISCARD,
-                    false, false, false, false);
-            MemberId member_id_struct_data = 0x00000000;
-            bool common_struct_data_ec {false};
-            CommonStructMember common_struct_data {TypeObjectUtils::build_common_struct_member(member_id_struct_data, member_flags_struct_data, TypeObjectUtils::retrieve_complete_type_identifier(type_ids_struct_data, common_struct_data_ec))};
-            if (!common_struct_data_ec)
-            {
-                EPROSIMA_LOG_ERROR(XTYPES_TYPE_REPRESENTATION, "Structure struct_data member TypeIdentifier inconsistent.");
-                return;
-            }
-            MemberName name_struct_data = "struct_data";
-            eprosima::fastcdr::optional<AppliedBuiltinMemberAnnotations> member_ann_builtin_struct_data;
-            ann_custom_StructType.reset();
-            CompleteMemberDetail detail_struct_data = TypeObjectUtils::build_complete_member_detail(name_struct_data, member_ann_builtin_struct_data, ann_custom_StructType);
-            CompleteStructMember member_struct_data = TypeObjectUtils::build_complete_struct_member(common_struct_data, detail_struct_data);
-            TypeObjectUtils::add_complete_struct_member(member_seq_StructType, member_struct_data);
-        }
-        CompleteStructType struct_type_StructType = TypeObjectUtils::build_complete_struct_type(struct_flags_StructType, header_StructType, member_seq_StructType);
-        if (eprosima::fastdds::dds::RETCODE_BAD_PARAMETER ==
-                TypeObjectUtils::build_and_register_struct_type_object(struct_type_StructType, type_name_StructType.to_string(), type_ids_StructType))
-        {
-            EPROSIMA_LOG_ERROR(XTYPES_TYPE_REPRESENTATION,
-                    "TrialModule::StructModule::StructType already registered in TypeObjectRegistry for a different type.");
-        }
-    }
-}
-
-} // namespace StructModule
-
-} // namespace TrialModule
-// TypeIdentifier is returned by reference: dependent structures/unions are registered in this same method
-void register_TrialStruct_type_identifier(
-        TypeIdentifierPair& type_ids_TrialStruct)
-{
-
-    ReturnCode_t return_code_TrialStruct {eprosima::fastdds::dds::RETCODE_OK};
-    return_code_TrialStruct =
-        eprosima::fastdds::dds::DomainParticipantFactory::get_instance()->type_object_registry().get_type_identifiers(
-        "TrialStruct", type_ids_TrialStruct);
-    if (eprosima::fastdds::dds::RETCODE_OK != return_code_TrialStruct)
-    {
-        StructTypeFlag struct_flags_TrialStruct = TypeObjectUtils::build_struct_type_flag(eprosima::fastdds::dds::xtypes::ExtensibilityKind::APPENDABLE,
-                false, false);
-        QualifiedTypeName type_name_TrialStruct = "TrialStruct";
-        eprosima::fastcdr::optional<AppliedBuiltinTypeAnnotations> type_ann_builtin_TrialStruct;
-        eprosima::fastcdr::optional<AppliedAnnotationSeq> ann_custom_TrialStruct;
-        CompleteTypeDetail detail_TrialStruct = TypeObjectUtils::build_complete_type_detail(type_ann_builtin_TrialStruct, ann_custom_TrialStruct, type_name_TrialStruct.to_string());
-        CompleteStructHeader header_TrialStruct;
-        header_TrialStruct = TypeObjectUtils::build_complete_struct_header(TypeIdentifier(), detail_TrialStruct);
-        CompleteStructMemberSeq member_seq_TrialStruct;
-        {
-            TypeIdentifierPair type_ids_struct_member;
-            ReturnCode_t return_code_struct_member {eprosima::fastdds::dds::RETCODE_OK};
-            return_code_struct_member =
-                eprosima::fastdds::dds::DomainParticipantFactory::get_instance()->type_object_registry().get_type_identifiers(
-                "TrialModule::StructModule::StructType", type_ids_struct_member);
-
-            if (eprosima::fastdds::dds::RETCODE_OK != return_code_struct_member)
-            {
-                TrialModule::StructModule::register_StructType_type_identifier(type_ids_struct_member);
-            }
-            StructMemberFlag member_flags_struct_member = TypeObjectUtils::build_struct_member_flag(eprosima::fastdds::dds::xtypes::TryConstructFailAction::DISCARD,
-                    false, false, false, false);
-            MemberId member_id_struct_member = 0x00000000;
-            bool common_struct_member_ec {false};
-            CommonStructMember common_struct_member {TypeObjectUtils::build_common_struct_member(member_id_struct_member, member_flags_struct_member, TypeObjectUtils::retrieve_complete_type_identifier(type_ids_struct_member, common_struct_member_ec))};
-            if (!common_struct_member_ec)
-            {
-                EPROSIMA_LOG_ERROR(XTYPES_TYPE_REPRESENTATION, "Structure struct_member member TypeIdentifier inconsistent.");
-                return;
-            }
-            MemberName name_struct_member = "struct_member";
-            eprosima::fastcdr::optional<AppliedBuiltinMemberAnnotations> member_ann_builtin_struct_member;
-            ann_custom_TrialStruct.reset();
-            CompleteMemberDetail detail_struct_member = TypeObjectUtils::build_complete_member_detail(name_struct_member, member_ann_builtin_struct_member, ann_custom_TrialStruct);
-            CompleteStructMember member_struct_member = TypeObjectUtils::build_complete_struct_member(common_struct_member, detail_struct_member);
-            TypeObjectUtils::add_complete_struct_member(member_seq_TrialStruct, member_struct_member);
-        }
-        CompleteStructType struct_type_TrialStruct = TypeObjectUtils::build_complete_struct_type(struct_flags_TrialStruct, header_TrialStruct, member_seq_TrialStruct);
-        if (eprosima::fastdds::dds::RETCODE_BAD_PARAMETER ==
-                TypeObjectUtils::build_and_register_struct_type_object(struct_type_TrialStruct, type_name_TrialStruct.to_string(), type_ids_TrialStruct))
-        {
-            EPROSIMA_LOG_ERROR(XTYPES_TYPE_REPRESENTATION,
-                    "TrialStruct already registered in TypeObjectRegistry for a different type.");
-        }
-    }
-}
-
