--- test/unittest/dds/xtypes/serializers/idl/types/module_idls/struct_struct/struct_struct.idl
@@ -1,15 +1,0 @@
-module TrialModule
-{
-    module StructModule
-    {
-        struct StructType
-        {
-            String struct_data;
-        };
-    };
-};
-    
-struct TrialStruct
-{
-    TrialModule::StructModule::StructType struct_member;
-};
