--- test/unittest/dds/xtypes/serializers/idl/types/module_idls/union_struct/union_struct.idl
@@ -1,20 +1,0 @@
-module TrialModule
-{
-    module UnionModule
-    {
-        union UnionType switch (long)
-        {
-            case 0:
-                long case_long;
-            case 1:
-                float case_float;
-            default:
-                string case_string;
-        };
-    };
-};
-
-struct TrialStruct
-{
-    TrialModule::UnionModule::UnionType union_member;
-};
