+++ test/unittest/dds/xtypes/serializers/idl/types/module_struct/gen/module_struct.hpp
@@ -13,14 +13,14 @@
 // limitations under the License.
 
 /*!
- * @file union_struct.hpp
+ * @file module_struct.hpp
  * This header file contains the declaration of the described types in the IDL file.
  *
  * This file was generated by the tool fastddsgen.
  */
 
-#ifndef FAST_DDS_GENERATED__UNION_STRUCT_HPP
-#define FAST_DDS_GENERATED__UNION_STRUCT_HPP
+#ifndef FAST_DDS_GENERATED__MODULE_STRUCT_HPP
+#define FAST_DDS_GENERATED__MODULE_STRUCT_HPP
 
 #include <cstdint>
 #include <functional>
@@ -42,25 +42,256 @@
 
 #if defined(_WIN32)
 #if defined(EPROSIMA_USER_DLL_EXPORT)
-#if defined(UNION_STRUCT_SOURCE)
-#define UNION_STRUCT_DllAPI __declspec( dllexport )
+#if defined(MODULE_STRUCT_SOURCE)
+#define MODULE_STRUCT_DllAPI __declspec( dllexport )
 #else
-#define UNION_STRUCT_DllAPI __declspec( dllimport )
-#endif // UNION_STRUCT_SOURCE
+#define MODULE_STRUCT_DllAPI __declspec( dllimport )
+#endif // MODULE_STRUCT_SOURCE
 #else
-#define UNION_STRUCT_DllAPI
+#define MODULE_STRUCT_DllAPI
 #endif  // EPROSIMA_USER_DLL_EXPORT
 #else
-#define UNION_STRUCT_DllAPI
+#define MODULE_STRUCT_DllAPI
 #endif // _WIN32
 
+namespace TrialModule {
+
+namespace AliasModule {
+
+typedef int32_t AliasType;
+
+
+} // namespace AliasModule
+
+} // namespace TrialModule
+namespace TrialModule {
+
+namespace BitsetModule {
+
+/*!
+ * @brief This structure represents the bitset BitsetType defined by the user in the IDL file.
+ * @ingroup module_struct
+ */
+struct BitsetType
+{
+        uint8_t flag1 : 3;
+
+        uint16_t flag2 : 10;
+
+        int32_t flag3 : 12;
+
+
+    /*!
+     * @brief Comparison operator.
+     * @param x BitsetType object to compare.
+     */
+    eProsima_user_DllExport bool operator ==(
+            const BitsetType& x) const
+    {
+        return (flag1 == x.flag1 &&
+           flag2 == x.flag2 &&
+           flag3 == x.flag3);
+    }
+
+    /*!
+     * @brief Comparison operator.
+     * @param x BitsetType object to compare.
+     */
+    eProsima_user_DllExport bool operator !=(
+            const BitsetType& x) const
+    {
+        return !(*this == x);
+    }
+};
+
+} // namespace BitsetModule
+
+} // namespace TrialModule
+namespace TrialModule {
+
+namespace BitmaskModule {
+
+/*!
+ * @brief This enumeration represents the BitmaskType bitflags defined by the user in the IDL file.
+ * @ingroup module_struct
+ */
+enum BitmaskTypeBits : uint8_t
+{
+    FLAG_ONE = 0x01ull << 0,
+    FLAG_TWO = 0x01ull << 1,
+    FLAG_THREE = 0x01ull << 2
+};
+typedef uint8_t BitmaskType;
+
+} // namespace BitmaskModule
+
+} // namespace TrialModule
+namespace TrialModule {
+
+namespace EnumModule {
+
+/*!
+ * @brief This class represents the enumeration EnumType defined by the user in the IDL file.
+ * @ingroup module_struct
+ */
+enum class EnumType : int32_t
+{
+    VALUE_ONE,
+    VALUE_TWO,
+    VALUE_THREE
+};
+
+} // namespace EnumModule
+
+} // namespace TrialModule
+namespace TrialModule {
+
+namespace StructModule {
+
+/*!
+ * @brief This class represents the structure StructType defined by the user in the IDL file.
+ * @ingroup module_struct
+ */
+class StructType
+{
+public:
+
+    /*!
+     * @brief Default constructor.
+     */
+    eProsima_user_DllExport StructType()
+    {
+    }
+
+    /*!
+     * @brief Default destructor.
+     */
+    eProsima_user_DllExport ~StructType()
+    {
+    }
+
+    /*!
+     * @brief Copy constructor.
+     * @param x Reference to the object StructType that will be copied.
+     */
+    eProsima_user_DllExport StructType(
+            const StructType& x)
+    {
+                    m_struct_data = x.m_struct_data;
+
+    }
+
+    /*!
+     * @brief Move constructor.
+     * @param x Reference to the object StructType that will be copied.
+     */
+    eProsima_user_DllExport StructType(
+            StructType&& x) noexcept
+    {
+        m_struct_data = std::move(x.m_struct_data);
+    }
+
+    /*!
+     * @brief Copy assignment.
+     * @param x Reference to the object StructType that will be copied.
+     */
+    eProsima_user_DllExport StructType& operator =(
+            const StructType& x)
+    {
+
+                    m_struct_data = x.m_struct_data;
+
+        return *this;
+    }
+
+    /*!
+     * @brief Move assignment.
+     * @param x Reference to the object StructType that will be copied.
+     */
+    eProsima_user_DllExport StructType& operator =(
+            StructType&& x) noexcept
+    {
+
+        m_struct_data = std::move(x.m_struct_data);
+        return *this;
+    }
+
+    /*!
+     * @brief Comparison operator.
+     * @param x StructType object to compare.
+     */
+    eProsima_user_DllExport bool operator ==(
+            const StructType& x) const
+    {
+        return (m_struct_data == x.m_struct_data);
+    }
+
+    /*!
+     * @brief Comparison operator.
+     * @param x StructType object to compare.
+     */
+    eProsima_user_DllExport bool operator !=(
+            const StructType& x) const
+    {
+        return !(*this == x);
+    }
+
+    /*!
+     * @brief This function copies the value in member struct_data
+     * @param _struct_data New value to be copied in member struct_data
+     */
+    eProsima_user_DllExport void struct_data(
+            const std::string& _struct_data)
+    {
+        m_struct_data = _struct_data;
+    }
+
+    /*!
+     * @brief This function moves the value in member struct_data
+     * @param _struct_data New value to be moved in member struct_data
+     */
+    eProsima_user_DllExport void struct_data(
+            std::string&& _struct_data)
+    {
+        m_struct_data = std::move(_struct_data);
+    }
+
+    /*!
+     * @brief This function returns a constant reference to member struct_data
+     * @return Constant reference to member struct_data
+     */
+    eProsima_user_DllExport const std::string& struct_data() const
+    {
+        return m_struct_data;
+    }
+
+    /*!
+     * @brief This function returns a reference to member struct_data
+     * @return Reference to member struct_data
+     */
+    eProsima_user_DllExport std::string& struct_data()
+    {
+        return m_struct_data;
+    }
+
+
+
+private:
+
+    std::string m_struct_data;
+
+};
+
+} // namespace StructModule
+
+} // namespace TrialModule
 namespace TrialModule {
 
 namespace UnionModule {
 
 /*!
  * @brief This class represents the union UnionType defined by the user in the IDL file.
- * @ingroup union_struct
+ * @ingroup module_struct
  */
 class UnionType
 {
@@ -518,55 +749,80 @@ private:
 
 } // namespace TrialModule
 /*!
- * @brief This class represents the structure TrialStruct defined by the user in the IDL file.
- * @ingroup union_struct
+ * @brief This class represents the structure ModuleStruct defined by the user in the IDL file.
+ * @ingroup module_struct
  */
-class TrialStruct
+class ModuleStruct
 {
 public:
 
     /*!
      * @brief Default constructor.
      */
-    eProsima_user_DllExport TrialStruct()
+    eProsima_user_DllExport ModuleStruct()
     {
     }
 
     /*!
      * @brief Default destructor.
      */
-    eProsima_user_DllExport ~TrialStruct()
+    eProsima_user_DllExport ~ModuleStruct()
     {
     }
 
     /*!
      * @brief Copy constructor.
-     * @param x Reference to the object TrialStruct that will be copied.
+     * @param x Reference to the object ModuleStruct that will be copied.
      */
-    eProsima_user_DllExport TrialStruct(
-            const TrialStruct& x)
+    eProsima_user_DllExport ModuleStruct(
+            const ModuleStruct& x)
     {
+                    m_alias_member = x.m_alias_member;
+
+                    m_bitset_member = x.m_bitset_member;
+
+                    m_bitmask_member = x.m_bitmask_member;
+
+                    m_enum_member = x.m_enum_member;
+
+                    m_struct_member = x.m_struct_member;
+
                     m_union_member = x.m_union_member;
 
     }
 
     /*!
      * @brief Move constructor.
-     * @param x Reference to the object TrialStruct that will be copied.
+     * @param x Reference to the object ModuleStruct that will be copied.
      */
-    eProsima_user_DllExport TrialStruct(
-            TrialStruct&& x) noexcept
+    eProsima_user_DllExport ModuleStruct(
+            ModuleStruct&& x) noexcept
     {
+        m_alias_member = x.m_alias_member;
+        m_bitset_member = std::move(x.m_bitset_member);
+        m_bitmask_member = std::move(x.m_bitmask_member);
+        m_enum_member = x.m_enum_member;
+        m_struct_member = std::move(x.m_struct_member);
         m_union_member = std::move(x.m_union_member);
     }
 
     /*!
      * @brief Copy assignment.
-     * @param x Reference to the object TrialStruct that will be copied.
+     * @param x Reference to the object ModuleStruct that will be copied.
      */
-    eProsima_user_DllExport TrialStruct& operator =(
-            const TrialStruct& x)
+    eProsima_user_DllExport ModuleStruct& operator =(
+            const ModuleStruct& x)
     {
+
+                    m_alias_member = x.m_alias_member;
+
+                    m_bitset_member = x.m_bitset_member;
+
+                    m_bitmask_member = x.m_bitmask_member;
+
+                    m_enum_member = x.m_enum_member;
+
+                    m_struct_member = x.m_struct_member;
 
                     m_union_member = x.m_union_member;
 
@@ -575,35 +831,220 @@ public:
 
     /*!
      * @brief Move assignment.
-     * @param x Reference to the object TrialStruct that will be copied.
+     * @param x Reference to the object ModuleStruct that will be copied.
      */
-    eProsima_user_DllExport TrialStruct& operator =(
-            TrialStruct&& x) noexcept
+    eProsima_user_DllExport ModuleStruct& operator =(
+            ModuleStruct&& x) noexcept
     {
 
+        m_alias_member = x.m_alias_member;
+        m_bitset_member = std::move(x.m_bitset_member);
+        m_bitmask_member = std::move(x.m_bitmask_member);
+        m_enum_member = x.m_enum_member;
+        m_struct_member = std::move(x.m_struct_member);
         m_union_member = std::move(x.m_union_member);
         return *this;
     }
 
     /*!
      * @brief Comparison operator.
-     * @param x TrialStruct object to compare.
+     * @param x ModuleStruct object to compare.
      */
     eProsima_user_DllExport bool operator ==(
-            const TrialStruct& x) const
+            const ModuleStruct& x) const
     {
-        return (m_union_member == x.m_union_member);
+        return (m_alias_member == x.m_alias_member &&
+           m_bitset_member == x.m_bitset_member &&
+           m_bitmask_member == x.m_bitmask_member &&
+           m_enum_member == x.m_enum_member &&
+           m_struct_member == x.m_struct_member &&
+           m_union_member == x.m_union_member);
     }
 
     /*!
      * @brief Comparison operator.
-     * @param x TrialStruct object to compare.
+     * @param x ModuleStruct object to compare.
      */
     eProsima_user_DllExport bool operator !=(
-            const TrialStruct& x) const
+            const ModuleStruct& x) const
     {
         return !(*this == x);
     }
+
+    /*!
+     * @brief This function sets a value in member alias_member
+     * @param _alias_member New value for member alias_member
+     */
+    eProsima_user_DllExport void alias_member(
+            TrialModule::AliasModule::AliasType _alias_member)
+    {
+        m_alias_member = _alias_member;
+    }
+
+    /*!
+     * @brief This function returns the value of member alias_member
+     * @return Value of member alias_member
+     */
+    eProsima_user_DllExport TrialModule::AliasModule::AliasType alias_member() const
+    {
+        return m_alias_member;
+    }
+
+    /*!
+     * @brief This function returns a reference to member alias_member
+     * @return Reference to member alias_member
+     */
+    eProsima_user_DllExport TrialModule::AliasModule::AliasType& alias_member()
+    {
+        return m_alias_member;
+    }
+
+
+    /*!
+     * @brief This function copies the value in member bitset_member
+     * @param _bitset_member New value to be copied in member bitset_member
+     */
+    eProsima_user_DllExport void bitset_member(
+            const TrialModule::BitsetModule::BitsetType& _bitset_member)
+    {
+        m_bitset_member = _bitset_member;
+    }
+
+    /*!
+     * @brief This function moves the value in member bitset_member
+     * @param _bitset_member New value to be moved in member bitset_member
+     */
+    eProsima_user_DllExport void bitset_member(
+            TrialModule::BitsetModule::BitsetType&& _bitset_member)
+    {
+        m_bitset_member = std::move(_bitset_member);
+    }
+
+    /*!
+     * @brief This function returns a constant reference to member bitset_member
+     * @return Constant reference to member bitset_member
+     */
+    eProsima_user_DllExport const TrialModule::BitsetModule::BitsetType& bitset_member() const
+    {
+        return m_bitset_member;
+    }
+
+    /*!
+     * @brief This function returns a reference to member bitset_member
+     * @return Reference to member bitset_member
+     */
+    eProsima_user_DllExport TrialModule::BitsetModule::BitsetType& bitset_member()
+    {
+        return m_bitset_member;
+    }
+
+
+    /*!
+     * @brief This function copies the value in member bitmask_member
+     * @param _bitmask_member New value to be copied in member bitmask_member
+     */
+    eProsima_user_DllExport void bitmask_member(
+            const TrialModule::BitmaskModule::BitmaskType& _bitmask_member)
+    {
+        m_bitmask_member = _bitmask_member;
+    }
+
+    /*!
+     * @brief This function moves the value in member bitmask_member
+     * @param _bitmask_member New value to be moved in member bitmask_member
+     */
+    eProsima_user_DllExport void bitmask_member(
+            TrialModule::BitmaskModule::BitmaskType&& _bitmask_member)
+    {
+        m_bitmask_member = std::move(_bitmask_member);
+    }
+
+    /*!
+     * @brief This function returns a constant reference to member bitmask_member
+     * @return Constant reference to member bitmask_member
+     */
+    eProsima_user_DllExport const TrialModule::BitmaskModule::BitmaskType& bitmask_member() const
+    {
+        return m_bitmask_member;
+    }
+
+    /*!
+     * @brief This function returns a reference to member bitmask_member
+     * @return Reference to member bitmask_member
+     */
+    eProsima_user_DllExport TrialModule::BitmaskModule::BitmaskType& bitmask_member()
+    {
+        return m_bitmask_member;
+    }
+
+
+    /*!
+     * @brief This function sets a value in member enum_member
+     * @param _enum_member New value for member enum_member
+     */
+    eProsima_user_DllExport void enum_member(
+            TrialModule::EnumModule::EnumType _enum_member)
+    {
+        m_enum_member = _enum_member;
+    }
+
+    /*!
+     * @brief This function returns the value of member enum_member
+     * @return Value of member enum_member
+     */
+    eProsima_user_DllExport TrialModule::EnumModule::EnumType enum_member() const
+    {
+        return m_enum_member;
+    }
+
+    /*!
+     * @brief This function returns a reference to member enum_member
+     * @return Reference to member enum_member
+     */
+    eProsima_user_DllExport TrialModule::EnumModule::EnumType& enum_member()
+    {
+        return m_enum_member;
+    }
+
+
+    /*!
+     * @brief This function copies the value in member struct_member
+     * @param _struct_member New value to be copied in member struct_member
+     */
+    eProsima_user_DllExport void struct_member(
+            const TrialModule::StructModule::StructType& _struct_member)
+    {
+        m_struct_member = _struct_member;
+    }
+
+    /*!
+     * @brief This function moves the value in member struct_member
+     * @param _struct_member New value to be moved in member struct_member
+     */
+    eProsima_user_DllExport void struct_member(
+            TrialModule::StructModule::StructType&& _struct_member)
+    {
+        m_struct_member = std::move(_struct_member);
+    }
+
+    /*!
+     * @brief This function returns a constant reference to member struct_member
+     * @return Constant reference to member struct_member
+     */
+    eProsima_user_DllExport const TrialModule::StructModule::StructType& struct_member() const
+    {
+        return m_struct_member;
+    }
+
+    /*!
+     * @brief This function returns a reference to member struct_member
+     * @return Reference to member struct_member
+     */
+    eProsima_user_DllExport TrialModule::StructModule::StructType& struct_member()
+    {
+        return m_struct_member;
+    }
+
 
     /*!
      * @brief This function copies the value in member union_member
@@ -647,10 +1088,15 @@ public:
 
 private:
 
+    TrialModule::AliasModule::AliasType m_alias_member{0};
+    TrialModule::BitsetModule::BitsetType m_bitset_member{};
+    TrialModule::BitmaskModule::BitmaskType m_bitmask_member{0};
+    TrialModule::EnumModule::EnumType m_enum_member{TrialModule::EnumModule::EnumType::VALUE_ONE};
+    TrialModule::StructModule::StructType m_struct_member;
     TrialModule::UnionModule::UnionType m_union_member;
 
 };
 
-#endif // _FAST_DDS_GENERATED_UNION_STRUCT_HPP_
+#endif // _FAST_DDS_GENERATED_MODULE_STRUCT_HPP_
