+++ test/unittest/dds/xtypes/serializers/idl/types/module_struct/gen/module_structCdrAux.hpp
@@ -13,19 +13,26 @@
 // limitations under the License.
 
 /*!
- * @file alias_structCdrAux.hpp
+ * @file module_structCdrAux.hpp
  * This source file contains some definitions of CDR related functions.
  *
  * This file was generated by the tool fastddsgen.
  */
 
-#ifndef FAST_DDS_GENERATED__ALIAS_STRUCTCDRAUX_HPP
-#define FAST_DDS_GENERATED__ALIAS_STRUCTCDRAUX_HPP
+#ifndef FAST_DDS_GENERATED__MODULE_STRUCTCDRAUX_HPP
+#define FAST_DDS_GENERATED__MODULE_STRUCTCDRAUX_HPP
 
-#include "alias_struct.hpp"
+#include "module_struct.hpp"
 
-constexpr uint32_t TrialStruct_max_cdr_typesize {8UL};
-constexpr uint32_t TrialStruct_max_key_cdr_typesize {0UL};
+constexpr uint32_t TrialModule_StructModule_StructType_max_cdr_typesize {264UL};
+constexpr uint32_t TrialModule_StructModule_StructType_max_key_cdr_typesize {0UL};
+
+
+
+
+constexpr uint32_t ModuleStruct_max_cdr_typesize {552UL};
+constexpr uint32_t ModuleStruct_max_key_cdr_typesize {0UL};
+
 
 
 
@@ -37,11 +44,15 @@ class CdrSizeCalculator;
 
 eProsima_user_DllExport void serialize_key(
         eprosima::fastcdr::Cdr& scdr,
-        const TrialStruct& data);
+        const TrialModule::StructModule::StructType& data);
+
+eProsima_user_DllExport void serialize_key(
+        eprosima::fastcdr::Cdr& scdr,
+        const ModuleStruct& data);
 
 
 } // namespace fastcdr
 } // namespace eprosima
 
-#endif // FAST_DDS_GENERATED__ALIAS_STRUCTCDRAUX_HPP
+#endif // FAST_DDS_GENERATED__MODULE_STRUCTCDRAUX_HPP
