+++ test/unittest/dds/xtypes/serializers/idl/types/module_struct/gen/module_structPubSubTypes.cxx
@@ -13,23 +13,47 @@
 // limitations under the License.
 
 /*!
- * @file struct_structPubSubTypes.cpp
+ * @file module_structPubSubTypes.cpp
  * This header file contains the implementation of the serialization functions.
  *
  * This file was generated by the tool fastddsgen.
  */
 
-#include "struct_structPubSubTypes.hpp"
+#include "module_structPubSubTypes.hpp"
 
 #include <fastdds/dds/log/Log.hpp>
 #include <fastdds/rtps/common/CdrSerialization.hpp>
 
-#include "struct_structCdrAux.hpp"
-#include "struct_structTypeObjectSupport.hpp"
+#include "module_structCdrAux.hpp"
+#include "module_structTypeObjectSupport.hpp"
 
 using SerializedPayload_t = eprosima::fastdds::rtps::SerializedPayload_t;
 using InstanceHandle_t = eprosima::fastdds::rtps::InstanceHandle_t;
 using DataRepresentationId_t = eprosima::fastdds::dds::DataRepresentationId_t;
+
+namespace TrialModule {
+    namespace AliasModule {
+    } // namespace AliasModule
+
+} // namespace TrialModule
+
+namespace TrialModule {
+    namespace BitsetModule {
+    } // namespace BitsetModule
+
+} // namespace TrialModule
+
+namespace TrialModule {
+    namespace BitmaskModule {
+    } // namespace BitmaskModule
+
+} // namespace TrialModule
+
+namespace TrialModule {
+    namespace EnumModule {
+    } // namespace EnumModule
+
+} // namespace TrialModule
 
 namespace TrialModule {
     namespace StructModule {
@@ -219,19 +243,25 @@ namespace TrialModule {
 
 } // namespace TrialModule
 
-TrialStructPubSubType::TrialStructPubSubType()
+namespace TrialModule {
+    namespace UnionModule {
+    } // namespace UnionModule
+
+} // namespace TrialModule
+
+ModuleStructPubSubType::ModuleStructPubSubType()
 {
-    set_name("TrialStruct");
-    uint32_t type_size = TrialStruct_max_cdr_typesize;
+    set_name("ModuleStruct");
+    uint32_t type_size = ModuleStruct_max_cdr_typesize;
     type_size += static_cast<uint32_t>(eprosima::fastcdr::Cdr::alignment(type_size, 4)); /* possible submessage alignment */
     max_serialized_type_size = type_size + 4; /*encapsulation*/
     is_compute_key_provided = false;
-    uint32_t key_length = TrialStruct_max_key_cdr_typesize > 16 ? TrialStruct_max_key_cdr_typesize : 16;
+    uint32_t key_length = ModuleStruct_max_key_cdr_typesize > 16 ? ModuleStruct_max_key_cdr_typesize : 16;
     key_buffer_ = reinterpret_cast<unsigned char*>(malloc(key_length));
     memset(key_buffer_, 0, key_length);
 }
 
-TrialStructPubSubType::~TrialStructPubSubType()
+ModuleStructPubSubType::~ModuleStructPubSubType()
 {
     if (key_buffer_ != nullptr)
     {
@@ -239,12 +269,12 @@ TrialStructPubSubType::~TrialStructPubSubType()
     }
 }
 
-bool TrialStructPubSubType::serialize(
+bool ModuleStructPubSubType::serialize(
         const void* const data,
         SerializedPayload_t& payload,
         DataRepresentationId_t data_representation)
 {
-    const TrialStruct* p_type = static_cast<const TrialStruct*>(data);
+    const ModuleStruct* p_type = static_cast<const ModuleStruct*>(data);
 
     // Object that manages the raw buffer.
     eprosima::fastcdr::FastBuffer fastbuffer(reinterpret_cast<char*>(payload.data), payload.max_size);
@@ -276,14 +306,14 @@ bool TrialStructPubSubType::serialize(
     return true;
 }
 
-bool TrialStructPubSubType::deserialize(
+bool ModuleStructPubSubType::deserialize(
         SerializedPayload_t& payload,
         void* data)
 {
     try
     {
         // Convert DATA to pointer of your type
-        TrialStruct* p_type = static_cast<TrialStruct*>(data);
+        ModuleStruct* p_type = static_cast<ModuleStruct*>(data);
 
         // Object that manages the raw buffer.
         eprosima::fastcdr::FastBuffer fastbuffer(reinterpret_cast<char*>(payload.data), payload.length);
@@ -306,7 +336,7 @@ bool TrialStructPubSubType::deserialize(
     return true;
 }
 
-uint32_t TrialStructPubSubType::calculate_serialized_size(
+uint32_t ModuleStructPubSubType::calculate_serialized_size(
         const void* const data,
         DataRepresentationId_t data_representation)
 {
@@ -317,7 +347,7 @@ uint32_t TrialStructPubSubType::calculate_serialized_size(
             eprosima::fastcdr::CdrVersion::XCDRv1 :eprosima::fastcdr::CdrVersion::XCDRv2);
         size_t current_alignment {0};
         return static_cast<uint32_t>(calculator.calculate_serialized_size(
-                    *static_cast<const TrialStruct*>(data), current_alignment)) +
+                    *static_cast<const ModuleStruct*>(data), current_alignment)) +
                 4u /*encapsulation*/;
     }
     catch (eprosima::fastcdr::exception::Exception& /*exception*/)
@@ -326,18 +356,18 @@ uint32_t TrialStructPubSubType::calculate_serialized_size(
     }
 }
 
-void* TrialStructPubSubType::create_data()
+void* ModuleStructPubSubType::create_data()
 {
-    return reinterpret_cast<void*>(new TrialStruct());
+    return reinterpret_cast<void*>(new ModuleStruct());
 }
 
-void TrialStructPubSubType::delete_data(
+void ModuleStructPubSubType::delete_data(
         void* data)
 {
-    delete(reinterpret_cast<TrialStruct*>(data));
+    delete(reinterpret_cast<ModuleStruct*>(data));
 }
 
-bool TrialStructPubSubType::compute_key(
+bool ModuleStructPubSubType::compute_key(
         SerializedPayload_t& payload,
         InstanceHandle_t& handle,
         bool force_md5)
@@ -347,7 +377,7 @@ bool TrialStructPubSubType::compute_key(
         return false;
     }
 
-    TrialStruct data;
+    ModuleStruct data;
     if (deserialize(payload, static_cast<void*>(&data)))
     {
         return compute_key(static_cast<void*>(&data), handle, force_md5);
@@ -356,7 +386,7 @@ bool TrialStructPubSubType::compute_key(
     return false;
 }
 
-bool TrialStructPubSubType::compute_key(
+bool ModuleStructPubSubType::compute_key(
         const void* const data,
         InstanceHandle_t& handle,
         bool force_md5)
@@ -366,17 +396,17 @@ bool TrialStructPubSubType::compute_key(
         return false;
     }
 
-    const TrialStruct* p_type = static_cast<const TrialStruct*>(data);
+    const ModuleStruct* p_type = static_cast<const ModuleStruct*>(data);
 
     // Object that manages the raw buffer.
     eprosima::fastcdr::FastBuffer fastbuffer(reinterpret_cast<char*>(key_buffer_),
-            TrialStruct_max_key_cdr_typesize);
+            ModuleStruct_max_key_cdr_typesize);
 
     // Object that serializes the data.
     eprosima::fastcdr::Cdr ser(fastbuffer, eprosima::fastcdr::Cdr::BIG_ENDIANNESS, eprosima::fastcdr::CdrVersion::XCDRv2);
     ser.set_encoding_flag(eprosima::fastcdr::EncodingAlgorithmFlag::PLAIN_CDR2);
     eprosima::fastcdr::serialize_key(ser, *p_type);
-    if (force_md5 || TrialStruct_max_key_cdr_typesize > 16)
+    if (force_md5 || ModuleStruct_max_key_cdr_typesize > 16)
     {
         md5_.init();
         md5_.update(key_buffer_, static_cast<unsigned int>(ser.get_serialized_data_length()));
@@ -396,11 +426,11 @@ bool TrialStructPubSubType::compute_key(
     return true;
 }
 
-void TrialStructPubSubType::register_type_object_representation()
+void ModuleStructPubSubType::register_type_object_representation()
 {
-    register_TrialStruct_type_identifier(type_identifiers_);
+    register_ModuleStruct_type_identifier(type_identifiers_);
 }
 
 
 // Include auxiliary functions like for serializing/deserializing.
-#include "struct_structCdrAux.ipp"
+#include "module_structCdrAux.ipp"
