+++ test/unittest/dds/xtypes/serializers/idl/types/module_struct/gen/module_structPubSubTypes.hpp
@@ -13,15 +13,15 @@
 // limitations under the License.
 
 /*!
- * @file struct_structPubSubTypes.hpp
+ * @file module_structPubSubTypes.hpp
  * This header file contains the declaration of the serialization functions.
  *
  * This file was generated by the tool fastddsgen.
  */
 
 
-#ifndef FAST_DDS_GENERATED__STRUCT_STRUCT_PUBSUBTYPES_HPP
-#define FAST_DDS_GENERATED__STRUCT_STRUCT_PUBSUBTYPES_HPP
+#ifndef FAST_DDS_GENERATED__MODULE_STRUCT_PUBSUBTYPES_HPP
+#define FAST_DDS_GENERATED__MODULE_STRUCT_PUBSUBTYPES_HPP
 
 #include <fastdds/dds/core/policy/QosPolicies.hpp>
 #include <fastdds/dds/topic/TopicDataType.hpp>
@@ -29,14 +29,39 @@
 #include <fastdds/rtps/common/SerializedPayload.hpp>
 #include <fastdds/utils/md5.hpp>
 
-#include "struct_struct.hpp"
+#include "module_struct.hpp"
 
 
 #if !defined(FASTDDS_GEN_API_VER) || (FASTDDS_GEN_API_VER != 3)
 #error \
-    Generated struct_struct is not compatible with current installed Fast DDS. Please, regenerate it with fastddsgen.
+    Generated module_struct is not compatible with current installed Fast DDS. Please, regenerate it with fastddsgen.
 #endif  // FASTDDS_GEN_API_VER
 
+namespace TrialModule
+{
+    namespace AliasModule
+    {
+        typedef int32_t AliasType;
+    } // namespace AliasModule
+} // namespace TrialModule
+namespace TrialModule
+{
+    namespace BitsetModule
+    {
+    } // namespace BitsetModule
+} // namespace TrialModule
+namespace TrialModule
+{
+    namespace BitmaskModule
+    {
+    } // namespace BitmaskModule
+} // namespace TrialModule
+namespace TrialModule
+{
+    namespace EnumModule
+    {
+    } // namespace EnumModule
+} // namespace TrialModule
 namespace TrialModule
 {
     namespace StructModule
@@ -44,7 +69,7 @@ namespace TrialModule
 
         /*!
          * @brief This class represents the TopicDataType of the type StructType defined by the user in the IDL file.
-         * @ingroup struct_struct
+         * @ingroup module_struct
          */
         class StructTypePubSubType : public eprosima::fastdds::dds::TopicDataType
         {
@@ -124,20 +149,26 @@ namespace TrialModule
         };
     } // namespace StructModule
 } // namespace TrialModule
+namespace TrialModule
+{
+    namespace UnionModule
+    {
+    } // namespace UnionModule
+} // namespace TrialModule
 
 /*!
- * @brief This class represents the TopicDataType of the type TrialStruct defined by the user in the IDL file.
- * @ingroup struct_struct
+ * @brief This class represents the TopicDataType of the type ModuleStruct defined by the user in the IDL file.
+ * @ingroup module_struct
  */
-class TrialStructPubSubType : public eprosima::fastdds::dds::TopicDataType
+class ModuleStructPubSubType : public eprosima::fastdds::dds::TopicDataType
 {
 public:
 
-    typedef TrialStruct type;
+    typedef ModuleStruct type;
 
-    eProsima_user_DllExport TrialStructPubSubType();
+    eProsima_user_DllExport ModuleStructPubSubType();
 
-    eProsima_user_DllExport ~TrialStructPubSubType() override;
+    eProsima_user_DllExport ~ModuleStructPubSubType() override;
 
     eProsima_user_DllExport bool serialize(
             const void* const data,
@@ -206,5 +237,5 @@ private:
 
 };
 
-#endif // FAST_DDS_GENERATED__STRUCT_STRUCT_PUBSUBTYPES_HPP
+#endif // FAST_DDS_GENERATED__MODULE_STRUCT_PUBSUBTYPES_HPP
